pyproject.toml
[project]
name = "vantage-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.3",
    "neo4j>=5.28.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "pydantic-settings>=2.2.1",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.3",
    "plotly>=6.1.2",
]
